# Improve performance of number comparison functions