### What problem does this PR solve?

Issue Number: ref #5672

Problem Summary:

### What is changed and how it works?
using dynamic dispatch
port from https://github.com/ClickHouse/ClickHouse/pull/37399

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

### Result
```mysql
--before
mysql > select c_name, c_custkey, o_orderkey, o_orderdate, o_totalprice, sum(l_quantity) from CUSTOMER, ORDERS, LINEITEM where o_orderkey in (         select l_orderkey         from LINEITEM
  group by l_orderkey         having sum(l_quantity) > 314     )     and c_custkey = o_custkey     and o_orderkey = l_orderkey group by c_name, c_custkey, o_orderkey, o_orderdate, o_totalprice order by o_totalprice desc, o_orderdate limit 100;
+--------------------+-----------+------------+-------------+--------------+-----------------+
| c_name             | c_custkey | o_orderkey | o_orderdate | o_totalprice | sum(l_quantity) |
+--------------------+-----------+------------+-------------+--------------+-----------------+
| Customer#000128120 |    128120 |    4722021 | 1994-04-07  |    544089.09 |          323.00 |
| Customer#000144617 |    144617 |    3043270 | 1997-02-12  |    530604.44 |          317.00 |
| Customer#000066790 |     66790 |    2199712 | 1996-09-30  |    515531.82 |          327.00 |
| Customer#000015619 |     15619 |    3767271 | 1996-08-07  |    480083.96 |          318.00 |
| Customer#000147197 |    147197 |    1263015 | 1997-02-02  |    467149.67 |          320.00 |
| Customer#000117919 |    117919 |    2869152 | 1996-06-20  |    456815.92 |          317.00 |
| Customer#000126865 |    126865 |    4702759 | 1994-11-07  |    447606.65 |          320.00 |
| Customer#000036619 |     36619 |    4806726 | 1995-01-17  |    446704.09 |          328.00 |
| Customer#000119989 |    119989 |    1544643 | 1997-09-20  |    434568.25 |          320.00 |
+--------------------+-----------+------------+-------------+--------------+-----------------+
9 rows in set (4.40 sec)

-- after 
mysql > select c_name, c_custkey, o_orderkey, o_orderdate, o_totalprice, sum(l_quantity) from CUSTOMER, ORDERS, LINEITEM where o_orderkey in (         select l_orderkey         from LINEITEM
  group by l_orderkey         having sum(l_quantity) > 314     )     and c_custkey = o_custkey     and o_orderkey = l_orderkey group by c_name, c_custkey, o_orderkey, o_orderdate, o_totalprice order by o_totalprice desc, o_orderdate limit 100;
+--------------------+-----------+------------+-------------+--------------+-----------------+
| c_name             | c_custkey | o_orderkey | o_orderdate | o_totalprice | sum(l_quantity) |
+--------------------+-----------+------------+-------------+--------------+-----------------+
| Customer#000128120 |    128120 |    4722021 | 1994-04-07  |    544089.09 |          323.00 |
| Customer#000144617 |    144617 |    3043270 | 1997-02-12  |    530604.44 |          317.00 |
| Customer#000066790 |     66790 |    2199712 | 1996-09-30  |    515531.82 |          327.00 |
| Customer#000015619 |     15619 |    3767271 | 1996-08-07  |    480083.96 |          318.00 |
| Customer#000147197 |    147197 |    1263015 | 1997-02-02  |    467149.67 |          320.00 |
| Customer#000117919 |    117919 |    2869152 | 1996-06-20  |    456815.92 |          317.00 |
| Customer#000126865 |    126865 |    4702759 | 1994-11-07  |    447606.65 |          320.00 |
| Customer#000036619 |     36619 |    4806726 | 1995-01-17  |    446704.09 |          328.00 |
| Customer#000119989 |    119989 |    1544643 | 1997-09-20  |    434568.25 |          320.00 |
+--------------------+-----------+------------+-------------+--------------+-----------------+
9 rows in set (4.24 sec)
```

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```