@Lloyd-Pottiger 2022-08-23
```shell
Running ./dbms/bench_dbms
Run on (40 X 2399.94 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x20)
  L1 Instruction 32 KiB (x20)
  L2 Unified 256 KiB (x20)
  L3 Unified 25600 KiB (x2)
Load Average: 48.56, 24.48, 23.63
# before
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
NumberComparsionBench/constantEqVector/iterations:100        95193 ns        95195 ns          100
NumberComparsionBench/vectorEqVector/iterations:100         112950 ns       112955 ns          100
NumberComparsionBench/constantLessVector/iterations:100     104847 ns       104850 ns          100
NumberComparsionBench/vectorLessVector/iterations:100       112402 ns       112410 ns          100
# after
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
NumberComparsionBench/constantEqVector/iterations:100        60521 ns        60521 ns          100
NumberComparsionBench/vectorEqVector/iterations:100          63688 ns        63687 ns          100
NumberComparsionBench/constantLessVector/iterations:100      58405 ns        58405 ns          100
NumberComparsionBench/vectorLessVector/iterations:100        64027 ns        64028 ns          100
```