@Lloyd-Pottiger 2022-08-23
change the `row_num = 10000000`
```shell
# before
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
NumberComparsionBench/constantEqVector/iterations:100     10734751 ns     10734228 ns          100
NumberComparsionBench/vectorEqVector/iterations:100       19793874 ns     19791053 ns          100
NumberComparsionBench/constantLessVector/iterations:100   11959062 ns     11947725 ns          100
NumberComparsionBench/vectorLessVector/iterations:100     19011253 ns     19001708 ns          100
# after
--------------------------------------------------------------------------------------------------
Benchmark                                                        Time             CPU   Iterations
--------------------------------------------------------------------------------------------------
NumberComparsionBench/constantEqVector/iterations:100      8810097 ns      8810259 ns          100
NumberComparsionBench/vectorEqVector/iterations:100       16357388 ns     16312559 ns          100
NumberComparsionBench/constantLessVector/iterations:100    9277899 ns      9278057 ns          100
NumberComparsionBench/vectorLessVector/iterations:100     15475154 ns     15475493 ns          100
```